c7/pyproject.toml
[tool.poetry]
name = "course7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
flake8 = "^6.1.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
eventlet = "^0.33.3"
celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
telebot = "^0.0.5"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
